JL8-24 refactor:임포트 경로 수정
