+++ src/app/(auth)/login/_components/LoginForm.tsx
@@ -1,4 +1,5 @@
 'use client'
+import { Button } from '@/components/common'
 import InputBox from '@/components/common/InputBox'
 import { useAuth } from '@/hooks/useAuth'
 import { userStore } from '@/store/userSlice'
@@ -6,7 +7,6 @@ import type { User, Users } from '@/types/auth'
 import { supabase } from '@/utils/supabase/client'
 import { useRouter } from 'next/navigation'
 import { useState } from 'react'
-import Button from '../../../../components/common/Button'
 // 텍스트 필드 라 어쩌고 제어 컴포넌트로 같이 처리 제어컴포넌트 / 비제어컴포넌트
 // constants 정규식 뺴주기
 // zod sschema 찾아보기
